+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+### Changed
+- Rename `getHistory` to `paginate` and the paramater `fromId` to `cursor`.
+  This is a breaking change.
+
 ## 0.7.0 -- 2017-07-20
 ### Changed
 - Rename the constructor option `instance` to `instanceId`. This is a breaking